+++ src/Factory/ContentStructureFactory.php
@@ -10,11 +10,8 @@ use CodeRhapsodie\EzDataflowBundle\Model\ContentUpdateStructure;
 use eZ\Publish\API\Repository\ContentService;
 use eZ\Publish\API\Repository\Exceptions\NotFoundException;
 
-final class ContentStructureFactory
+final class ContentStructureFactory implements ContentStructureFactoryInterface
 {
-    public const MODE_INSERT_OR_UPDATE = 1;
-    public const MODE_INSERT_ONLY = 2;
-    public const MODE_UPDATE_ONLY = 3;
     /**
      * @var ContentService
      */
@@ -36,14 +33,14 @@ final class ContentStructureFactory
      * @param string $language
      * @param string $contentType
      * @param mixed  $parentLocations
-     * @param int    $mode            One of the constant ContentStructureFactory::MODE_*
+     * @param int    $mode            One of the constant ContentStructureFactoryInterface::MODE_*
      *
      * @return false|ContentStructure
      *
      * @throws \CodeRhapsodie\EzDataflowBundle\Exception\InvalidArgumentTypeException
      * @throws \eZ\Publish\API\Repository\Exceptions\UnauthorizedException
      */
-    public function transform(array $data, string $remoteId, string $language, string $contentType, $parentLocations, int $mode = ContentStructureFactory::MODE_INSERT_OR_UPDATE)
+    public function transform(array $data, string $remoteId, string $language, string $contentType, $parentLocations, int $mode = ContentStructureFactoryInterface::MODE_INSERT_OR_UPDATE)
     {
         try {
             $content = $this->contentService->loadContentByRemoteId($remoteId);