+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * Add compiler pass to change the Dataflow DBAL connection factory
 * Use the DBAL connection from siteaccess
 * Add `mode` argument on `ContentStructureFactory::transform()` function
+* Add `CodeRhapsodie\EzDataflowBundle\Factory\ContentStructureFactoryInterface`
 
 # version 1.0.0